MULE-8741: Extract logic for creating implicit configs to a separate class.
