+++ core/src/main/java/org/mule/config/builders/ExtensionsManagerConfigurationBuilder.java
@@ -6,9 +6,9 @@
  */
 package org.mule.config.builders;
 
+import static org.mule.api.lifecycle.LifecycleUtils.initialiseIfNeeded;
 import org.mule.DefaultMuleContext;
 import org.mule.api.MuleContext;
-import org.mule.api.context.MuleContextAware;
 import org.mule.extension.ExtensionManager;
 import org.mule.util.ClassUtils;
 
@@ -34,10 +34,7 @@ public class ExtensionsManagerConfigurationBuilder extends AbstractConfiguration
             extensionManager.discoverExtensions(Thread.currentThread().getContextClassLoader());
 
             ((DefaultMuleContext) muleContext).setExtensionManager(extensionManager);
-            if (extensionManager instanceof MuleContextAware)
-            {
-                ((MuleContextAware) extensionManager).setMuleContext(muleContext);
-            }
+            initialiseIfNeeded(extensionManager, muleContext);
         }
     }
 }
